rating-calculator/requirements.txt
python-dotenv==1.0.0
numpy>=1.24.4
pandas>=2.0.3
requests>=2.28.0